move every component reimport into Basic folder
