+++ src/components/App/app.scss
@@ -17,6 +17,7 @@
   }
 
   .content {
-    padding: 20px;
+    margin-left: 10px;
+    margin-right: 10px;
   }
 }
