+++ src/components/Basic/index.ts
@@ -1,0 +1,7 @@
+import View from './View';
+import ScrollView from './ScrollView'
+
+export {
+  View,
+  ScrollView
+};
